server/requirements.txt
django==4.1.7
djangorestframework==3.14
django-rest-knox==4.2.0
pillow==9.4.0
django-debug-toolbar==3.8.1
drf-nested-routers==0.10.0 
django-jazzmin==2.6.0
django-filter==22.1
drf-spectacular==0.26.0
gunicorn==20.1.0
django-cors-headers==3.14.0
